fix: Distinct between float and integer when printing YAML format (#89)
## Motivation

Since version v0.0.97 a regression has occurred where we stopped
printing integer values correctly. Instead of printing them as integers
we always printed every number as a float.
The main culprit of this regression was a switch to SDK-based parsing
methods which utilized a new YAML library.

## Related changes

Fixed in: https://github.com/nobl9/nobl9-go/pull/347.

## Testing

Run `sloctl get` on objects which operate on ints and floats like Alert
Policy or Agent.

## Release Notes

Fixed regression in printing YAML output for `sloctl get` commands which
caused integers to be displayed as floats (ex. `1.0` instead of `1`).
The regression was introduced in version `v0.0.97`.