chore: Queue/Trigger the replay using sloctl [PC-13394] (#237)
## Motivation

We're adding a new way to create a Replay. The new Replay is added to
the queue and there's no need to wait for the reimport to complete.

## Related changes

A new way of creating a Replay. Considering the number of ongoing
Replays is limited, `sloctl` queues Replays if the limit is exceeded.
Replay queues is an experimental feature, currently unavailable to all
organizations. We're working on improving and expanding its
availability.

---------

Co-authored-by: paulina-n9 <65724378+paulina-n9@users.noreply.github.com>