chore: Fix bats tests (#306)
## Motivation

With the changes introduced to `sloctl version` command in
https://github.com/nobl9/sloctl/pull/304
we no longer have the embedded sloctl version from `VERSION` file as
fallback.
Instead, we're inferring the version (if it was not set during build
time) using `runtime/debug` package.

This works fine, except for scenarios where we build sloctl binary
without the version.
One instance of such usage are bats e2e tests.
Due to that, and the fallback version being `0.0.0` the API was not
returning the AWS IAM id for data export as it only returns it for
sloctl version `v0.0.94` and higher.

## Summary

Version is now set for all Makefile build commands, it is inferred from
the current latest git tag and a build suffix is added, which will look
like this: `0.12.0-test`.

## Testing

Run the e2e tests.

Dispatched tests run:
https://github.com/nobl9/sloctl/actions/runs/13819778124/job/38662027117.
The run failed on another step cause I cancelled the run before this run
and it left dangling resources.