feat: Add cancellation field to replay list response (#332)
## Motivation

The ability to cancel replays when they are in the importing phase is in
development. New action will be available in both UI and sloctl.

## Summary

This pull request includes a renaming of a struct and the addition of a
new field to improve clarity and functionality in the replay list
feature.

### Struct renaming and enhancement:

* Renamed the `ReplayQueueItem` struct to `ReplayListItem` for better
alignment with its purpose in the replay list context.
(`internal/replay_list.go`,
[internal/replay_list.goL24-R31](diffhunk://#diff-73224e3f4249a9b4533174b18bdabf4cdbaf00f507d9116c78d48b9e40c27c57L24-R31))
* Added a new field `Cancellation` (of type `string`) to the
`ReplayListItem` struct to capture cancellation-related information.
(`internal/replay_list.go`,
[internal/replay_list.goL24-R31](diffhunk://#diff-73224e3f4249a9b4533174b18bdabf4cdbaf00f507d9116c78d48b9e40c27c57L24-R31))

### Code adjustments for the renamed struct:

* Updated the variable type in the `listAllReplays` function to use the
renamed `ReplayListItem` struct instead of the old `ReplayQueueItem`.
(`internal/replay_list.go`,
[internal/replay_list.goL48-R49](diffhunk://#diff-73224e3f4249a9b4533174b18bdabf4cdbaf00f507d9116c78d48b9e40c27c57L48-R49))

## Related changes

https://github.com/nobl9/n9/pull/17808

## Testing

```bash
sloctl replay list
```
```yaml
- slo: cancel-test-4
  project: prometheus
  elapsedTime: 2m0s
  retrievedScope: 13 Day
  retrievedFrom: "2025-04-30T23:55:34Z"
  status: in progress
  cancellation: possible
- slo: cancel-test
  project: prometheus
  elapsedTime: 2m0s
  retrievedScope: 13 Day
  retrievedFrom: "2025-04-30T23:55:28Z"
  status: in progress
  cancellation: possible
- slo: cancel-test-3
  project: prometheus
  elapsedTime: 2m0s
  retrievedScope: 13 Day
  retrievedFrom: "2025-04-30T23:55:32Z"
  status: queued
  cancellation: blocked
```

## Release Notes

The output of `sloctl replay list` now has a new `cancellation` field,
which contains information about the availability and status of the
cancellation of the replay.